Refactor Hash out of DB project for AB#13730
