+++ Apps/Common/src/Delegates/HmacHashDelegate.cs
@@ -17,9 +17,9 @@ namespace HealthGateway.Common.Delegates
 {
     using System;
     using System.Security.Cryptography;
+    using HealthGateway.Common.Constants;
     using HealthGateway.Common.Models;
-    using HealthGateway.Database.Constants;
-    using HealthGateway.Database.Models.Cacheable;
+    using HealthGateway.Common.Models.Cacheable;
     using Microsoft.AspNetCore.Cryptography.KeyDerivation;
     using Microsoft.Extensions.Configuration;
