+++ Apps/Database/src/Database.csproj
@@ -26,7 +26,6 @@
 
   <ItemGroup>
     <Folder Include="Wrapper\" />
-    <Folder Include="Models\Cacheable\" />
   </ItemGroup>
